requirements.txt
ffmpeg-python
moviepy
groq
nltk
google-genai
google
sentence-transformers
pydantic
python-dotenv
numpy
scipy
scikit-learn
torch
pillow
transformers
